pyproject.toml
[tool.poetry]
name = "book-search"
version = "1.0.0"
description = "API para busca de livros"
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pandas = "^2.3.1"
requests = "^2.32.4"
pydantic = "^2.11.7"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-dotenv = "^1.1.1"
beautifulsoup4 = "^4.13.4"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pre-commit = "^4.2.0"
ruff = "^0.12.7"
yapf = "^0.43.0"
isort = "^6.0.1"
pylint = "^3.3.7"

[tool.yapf]
based_on_style = "pep8"
column_limit = 140
split_complex_comprehension = true
coalesce_brackets = true
allow_multiline_dictionary_keys = true
align_closing_bracket_with_visual_indent = true
split_penalty_import_names = 1
dedent_closing_brackets = true

[tool.ruff]
line-length = 140
indent-width = 4
target-version = "py312"

[tool.ruff.lint]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = 140

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.isort]
line_length = 140
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = false
split_on_trailing_comma = true
force_alphabetical_sort_within_sections = true


[tool.pylint]
disable = "E1101, E1133"